54 order 주문 도메인 entity dto 비즈니스 로직 구현

+++ backend/src/main/java/com/example/pproject/application/repository/JobApplicationRepository.java
@@ -8,10 +8,10 @@ import java.util.List;
 public interface JobApplicationRepository extends JpaRepository<JobApplication, Long> {
 
     // 특정 유저의 지원 내역 조회 (최신순)
-    List<JobApplication> findByMemberIdOrderByAppliedAtDesc(Integer memberId);
+    List<JobApplication> findByMemberIdOrderByAppliedAtDesc(Long memberId);
 
     // 중복 지원 확인
-    boolean existsByJobIdAndMemberId(Long jobId, Integer memberId);
+    boolean existsByJobIdAndMemberId(Long jobId, Long memberId);
 
     // 이력서 삭제 시 참조 확인
     boolean existsByResumeId(Long resumeId);
